AmspecConnector/requirements.txt
requests
python-dotenv
httpx
urllib3
tls_client

